refactor daemon acceptance ownership
